Overhaul controller-client interactions, add startup package
- Removed redundant Client interface and replaced with concrete class
Client (previously named FXClient)
- Added startup package to house the serving as the runnable application
- Implemented SceneBuilder and LoginSceneBuilder (more work needs to be
done in handling events from this view. Added relevant TODO's in the
handler functions in controller)
- Revamped the use of template design pattern in SceneBuilder and
subclasses, changed content of created subclasses to fit this new
interface
- Modified Client constructor to take only a Stage as param. This means
that it is the responsibility of the controller to call
setController(this) on the client.
- Added this functionality in the controller constructor along with a
stack to hold ViewTypes, and a call to SetView(LOGIN) at construction.